+++ src/geneflow/cli/install_workflow.py
@@ -74,7 +74,7 @@ def init_subparser(subparsers):
         type=str,
         required=False,
         default=None,
-        help='GeneFlow config file with agave connection information'
+        help='GeneFlow config file'
     )
     parser.add_argument(
         '-e', '--environment',
@@ -83,84 +83,6 @@ def init_subparser(subparsers):
         default=None,
         help='config environment'
     )
-    parser.add_argument(
-        '--agave-params',
-        type=str,
-        required=False,
-        default=None,
-        dest='agave_params',
-        help='Agave params file for registration'
-    )
-    parser.add_argument(
-        '--agave-username',
-        type=str,
-        required=False,
-        default=None,
-        dest='agave_username',
-        help='Agave username to impersonate'
-    )
-    parser.add_argument(
-        '--agave-domain',
-        type=str,
-        required=False,
-        default='',
-        dest='agave_domain',
-        help='Agave domain for user store'
-    )
-    parser.add_argument(
-        '--agave-apps-prefix',
-        type=str,
-        required=False,
-        default=None,
-        dest='agave_apps_prefix',
-        help='App name prefix for Agave registration'
-    )
-    parser.add_argument(
-        '--agave-execution-system',
-        type=str,
-        required=False,
-        default=None,
-        dest='agave_execution_system',
-        help='Execution system for Agave registration'
-    )
-    parser.add_argument(
-        '--agave-deployment-system',
-        type=str,
-        required=False,
-        default=None,
-        dest='agave_deployment_system',
-        help='Deployment system for Agave registration'
-    )
-    parser.add_argument(
-        '--agave-apps-dir',
-        type=str,
-        required=False,
-        default=None,
-        dest='agave_apps_dir',
-        help='Apps directory for Agave registration'
-    )
-    parser.add_argument(
-        '--agave-test-data-dir',
-        type=str,
-        required=False,
-        default=None,
-        dest='agave_test_data_dir',
-        help='Test data directory for Agave registration'
-    )
-    parser.add_argument(
-        '--agave-publish', action='store_true',
-        required=False,
-        dest='agave_publish',
-        help='Publish Agave app'
-    )
-    parser.add_argument(
-        '--agave-test-data', action='store_true',
-        required=False,
-        dest='agave_test_data',
-        help='Upload Agave test data'
-    )
-    parser.set_defaults(agave_publish=False)
-    parser.set_defaults(agave_test_data=False)
     parser.set_defaults(func=install_workflow)
 
     return parser
@@ -202,43 +124,6 @@ def install_workflow(args, other_args, subparser=None):
         cfg.default('database.db')
         config_dict = cfg.config('local')
 
-    # load agave params if specified
-    agave_params = {}
-    if args.agave_params:
-        try:
-            with open(args.agave_params, 'rU') as yaml_file:
-                yaml_data = yaml_file.read()
-        except IOError as err:
-            Log.an().error(
-                'cannot read agave params file: %s [%s]',
-                args.params,
-                str(err)
-            )
-            return False
-
-        try:
-            agave_params = yaml.safe_load(yaml_data)
-        except yaml.YAMLError as err:
-            Log.an().error(
-                'invalid yaml: %s [%s]', yaml_data, str(err)
-            )
-            return False
-
-    if not agave_params.get('agave'):
-        agave_params['agave'] = {}
-
-    # override any agave_params keys with command line options
-    if args.agave_apps_prefix:
-        agave_params['agave']['appsPrefix'] = args.agave_apps_prefix
-    if args.agave_execution_system:
-        agave_params['agave']['executionSystem'] = args.agave_execution_system
-    if args.agave_deployment_system:
-        agave_params['agave']['deploymentSystem'] = args.agave_deployment_system
-    if args.agave_apps_dir:
-        agave_params['agave']['appsDir'] = args.agave_apps_dir
-    if args.agave_test_data_dir:
-        agave_params['agave']['testDataDir'] = args.agave_test_data_dir
-
     # initialize workflow installer object and install apps
     wf_installer = WorkflowInstaller(
         str(Path(args.workflow_path).resolve()),
@@ -248,10 +133,6 @@ def install_workflow(args, other_args, subparser=None):
         app_name=args.name,
         clean=args.clean,
         config=config_dict,
-        agave_params=agave_params,
-        agave_username=args.agave_username,
-        agave_domain=args.agave_domain,
-        agave_publish=args.agave_publish,
         make_apps=args.make_apps
     )
 
@@ -262,10 +143,5 @@ def install_workflow(args, other_args, subparser=None):
     if not wf_installer.install_apps():
         Log.an().error('cannot install workflow apps')
         return False
-
-    if args.agave_test_data:
-        if not wf_installer.upload_agave_test_data():
-            Log.an().error('cannot upload agave test data')
-            return False
 
     return True