+++ src/geneflow/data/sql/geneflow.sql
@@ -1,5 +1,3 @@
--- workflow schema 20200526-00
-
 drop table if exists workflow;
 drop table if exists app;
 drop table if exists step;
@@ -13,15 +11,12 @@ create table workflow (
     name varchar(256) not null,
     description varchar(256) not null default '',
     version varchar(32) not null default '',
-    username varchar(32) not null default '',
+    author varchar(32) not null default '',
     git text not null default '',
     inputs text not null default '',
     parameters text not null default '',
-    final_output text not null default '',
+    publish text not null default '',
     apps text not null default '',
-    public tinyint not null default 0,
-    enable tinyint not null default 1,
-    test tinyint not null default 0,
     created timestamp default '0000-00-00 00:00:00',
     modified timestamp default CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP,
     primary key (id)
@@ -33,14 +28,14 @@ create table app (
     description varchar(256) not null default '',
     git text not null default '',
     version varchar(32) not null default '',
-    username varchar(32) not null default '',
+    author varchar(32) not null default '',
     implementation text not null default '',
     inputs text not null default '',
     parameters text not null default '',
-    public tinyint not null default 0,
-    pre_exec text not null default '',
+    images text not null default '',
+    exec_pre text not null default '',
     exec_methods text not null default '',
-    post_exec text not null default '',
+    exec_post text not null default '',
     primary key (id)
 );
 
@@ -59,8 +54,11 @@ create table step (
     workflow_id char(32) not null,
     app_id char(32) not null,
     map_uri varchar(256) not null default '',
+    map_inclusive tinyint not null default 0,
+    map_glob varchar(256) not null default '*',
     map_regex varchar(256) not null default '',
     template text not null default '',
+    publish tinyint not null default 0,
     exec_context varchar(256) not null default 'local',
     exec_method varchar(256) not null default 'auto',
     exec_parameters text not null default '',
@@ -82,11 +80,10 @@ create table job (
     inputs text not null default '',
     parameters text not null default '',
     output_uri varchar(256) not null default '',
-    final_output text not null default '',
+    publish text not null default '',
     exec_context text not null default '',
     exec_method text not null default '',
     exec_parameters text not null default '',
-    notifications text not null default '[]',
     primary key (id)
 );