+++ src/geneflow/workflow_dag.py
@@ -35,7 +35,7 @@ class WorkflowDAG:
 
         Args:
             workflow: dict of workflow, updated to include any inputs,
-                parameters, and final_output specified in the job.
+                parameters, and publish steps specified in the job.
             apps: dict of apps:
                 {
                     'app1': { app dict },
@@ -47,7 +47,7 @@ class WorkflowDAG:
             parsed_job_output_uri: job output URI.
             config: dict of workflow config
             **kwargs: additional arguments specific to contexts for this
-                workflow (e.g., local or agave arguments)
+                workflow (e.g., local arguments)
 
         Returns:
             WorkflowDAG instance.
@@ -436,7 +436,7 @@ class WorkflowDAG:
         Create empty nodes for each workflow input and step.
 
         Nodes contain attributes for type (e.g., input or step), contexts for
-        data staging (e.g., local or agave), source context, and node.
+        data staging (e.g., local), source context, and node.
         The node attribute is initialized as None, but will later be a
         reference to a WorkflowInput or WorkflowStep object.
 
@@ -495,7 +495,7 @@ class WorkflowDAG:
                 raise WorkflowDAGException(msg)
 
             contexts = {source_data_context: ''}
-            if step_name in self._workflow['final_output']:
+            if step_name in self._workflow['publish']:
                 contexts['final'] = ''
 
             try: